Move drawing order login to helper class
